+++ src/rules/index.js
@@ -1,15 +1,15 @@
+import atExtendNoMissingPlaceholder from "./at-extend-no-missing-placeholder"
 import atFunctionPattern from "./at-function-pattern"
 import atMixinPattern from "./at-mixin-pattern"
 import dollarVariableNoMissingInterpolation from "./dollar-variable-no-missing-interpolation"
 import dollarVariablePattern from "./dollar-variable-pattern"
-import extendRequirePlaceholder from "./extend-require-placeholder"
 import selectorNoRedundantParent from "./selector-no-redundant-parent"
 
 export default {
+  "at-extend-no-missing-placeholder": atExtendNoMissingPlaceholder,
   "at-function-pattern": atFunctionPattern,
   "at-mixin-pattern": atMixinPattern,
   "dollar-variable-no-missing-interpolation": dollarVariableNoMissingInterpolation,
   "dollar-variable-pattern": dollarVariablePattern,
-  "extend-require-placeholder": extendRequirePlaceholder,
   "selector-no-redundant-parent": selectorNoRedundantParent,
 }