+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 # 1.0.0
 
+- Added: `at-extend-no-missing-placeholder` rule.
 - Added: `at-function-pattern` rule.
 - Added: `at-mixin-pattern` rule.
 - Added: `dollar-variable-no-missing-interpolation` rule.
 - Added: `dollar-variable-pattern` rule.
-- Added: `extend-require-placeholder` rule.
 - Added: `selector-no-redundant-parent` rule.